hydra_mlflow_example/requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
alembic==1.4.1
antlr4-python3-runtime==4.8
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
cloudpickle==1.6.0
databricks-cli==0.14.3
docker==5.0.0
entrypoints==0.3
Flask==2.0.0
fsspec==2021.4.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.30.0
google-auth-oauthlib==0.4.4
greenlet==1.1.0
grpcio==1.37.1
gunicorn==20.1.0
hydra-core==1.0.6
idna==2.10
importlib-resources==5.1.2
itsdangerous==2.0.0
Jinja2==3.0.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.0
mlflow==1.17.0
multidict==5.1.0
numpy==1.20.3
oauthlib==3.1.0
omegaconf==2.0.6
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
prometheus-client==0.10.1
prometheus-flask-exporter==0.18.2
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytorch-lightning==1.3.1
pytz==2021.1
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.15
sqlparse==0.4.1
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
torch==1.8.1
torchmetrics==0.3.2
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
utils==1.0.1
websocket-client==0.59.0
Werkzeug==2.0.0
yarl==1.6.3
